// ui/src/services/cradle/models/DigestUploadFinalizeCreateRequest.ts
// @ts-nocheck
/* tslint:disable */
/* eslint-disable */
/**
 * CRADLE
 * Threat Intelligence Knowledge Management
 *
 * The version of the OpenAPI document: 2.10.2-beta.9bd46310 (2.10.2-beta.9bd46310)
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * Serializer for finalizing digest upload (creates digest metadata only).
 * 
 * The digest file is uploaded directly to storage via presigned URL and must exist
 * at BaseDigest.storage_key before finalization.
 * @export
 * @interface DigestUploadFinalizeCreateRequest
 */
export interface DigestUploadFinalizeCreateRequest {
    /**
     * 
     * @type {string}
     * @memberof DigestUploadFinalizeCreateRequest
     */
    title: string;
    /**
     * 
     * @type {string}
     * @memberof DigestUploadFinalizeCreateRequest
     */
    digestType: string;
    /**
     * Optional entities to associate with this digest
     * @type {Array<number>}
     * @memberof DigestUploadFinalizeCreateRequest
     */
    entities?: Array<number>;
}

/**
 * Check if a given object implements the DigestUploadFinalizeCreateRequest interface.
 */
export function instanceOfDigestUploadFinalizeCreateRequest(value: object): value is DigestUploadFinalizeCreateRequest {
    if (!('title' in value) || value['title'] === undefined) return false;
    if (!('digestType' in value) || value['digestType'] === undefined) return false;
    return true;
}

export function DigestUploadFinalizeCreateRequestFromJSON(json: any): DigestUploadFinalizeCreateRequest {
    return DigestUploadFinalizeCreateRequestFromJSONTyped(json, false);
}

export function DigestUploadFinalizeCreateRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): DigestUploadFinalizeCreateRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'title': json['title'],
        'digestType': json['digest_type'],
        'entities': json['entities'] == null ? undefined : json['entities'],
    };
}

export function DigestUploadFinalizeCreateRequestToJSON(json: any): DigestUploadFinalizeCreateRequest {
    return DigestUploadFinalizeCreateRequestToJSONTyped(json, false);
}

export function DigestUploadFinalizeCreateRequestToJSONTyped(value?: DigestUploadFinalizeCreateRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'title': value['title'],
        'digest_type': value['digestType'],
        'entities': value['entities'],
    };
}